Moved HelpText styles to CSS and styling cleanup
