+++ autoscheduler/frontend/src/components/App/App.css
@@ -1,7 +1,9 @@
 .app-container {
     height: 100%;
+    display: flex;
+    flex-direction: column;
 }
 
 .app-container > div:nth-child(2) {
-    height: calc(100% - 64px);
+    flex-grow: 1;
 }
