+++ autoscheduler/frontend/src/components/LandingPage/HelpText/HelpText.css
@@ -1,0 +1,7 @@
+.container {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    align-items: center;
+    margin-bottom: 2%;
+}
