+++ autoscheduler/frontend/src/components/LandingPage/LandingPage.css
@@ -1,0 +1,6 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    justify-content: center;
+}
